Move common search_hits functionality to its own module
